-Ticket #2936
git-svn-id: https://subversion.assembla.com/svn/mdss@7444 e0e46c49-be69-4f5a-ad62-21024a331aea

Former-commit-id: 052b027a6aadad8cdd48a3627e963ce1fe8ae35a [formerly f610263db2b04cae79a645b5d71aa7c7e9ced074]
Former-commit-id: eae17fa2bc71b3951a3590e7f0b0549f87ff53d2

+++ trunk/src/common/dss/vector/solutions/report/CacheDocumentManager.java
@@ -51,13 +51,16 @@ public class CacheDocumentManager implements Runnable
 
         File[] files = directory.listFiles();
 
-        for (File file : files)
+        if (files != null)
         {
-          String sessionId = new String(Base64.decode(file.getName()));
-
-          if (!SessionFacade.containsSession(sessionId))
+          for (File file : files)
           {
-            FileIO.deleteDirectory(file);
+            String sessionId = new String(Base64.decode(file.getName()));
+
+            if (!SessionFacade.containsSession(sessionId))
+            {
+              FileIO.deleteDirectory(file);
+            }
           }
         }
       }
